Rename shared/ → common/ (Adam's suggestion)
